move find_object_in_stack helper to utils module.
